Rename {Default => Configured}SslContextFactoryProvider, {Throwing => Default}SslContextFactoryProvider
Move SslContextFactoryProvider implementations to 'impl' package.
